refactor: delete kakao share container and integrate to modal
