+++ packages/web-domains/src/common/components/KakaoShare/KakaoShareModal.tsx
@@ -1,11 +1,17 @@
 'use client';
 
-import { HTMLAttributes, useEffect } from 'react';
+import { Button, Txt } from '@sds/components';
+import { colors } from '@sds/theme';
+import { HTMLAttributes, useEffect, useState } from 'react';
 
 import { Modal } from '../Modal/Modal';
 
+import { CopyLinkButton } from './CopyLinkButton';
+import { copyToClipboard } from './copyToClipboard.utils';
 import { KakaoShareOpenGraphKeyType } from './generateKakaoShare.utils';
-import { KakaoShareContainer } from './KakaoShare.container';
+import { ShareKakaoButton } from './KakaoShareButton';
+import { buttonWrapperCss, textWrapperCss } from './KakaoShareModal.styles';
+import { Toast } from './Toast';
 
 interface ModalProps extends HTMLAttributes<HTMLDivElement> {
   isOpen: boolean;
@@ -15,6 +21,18 @@ interface ModalProps extends HTMLAttributes<HTMLDivElement> {
 }
 
 export const KakaoShareModal = ({ title, openGraphKey, isOpen, onClose, ...rest }: ModalProps) => {
+  const [isToastOpen, setIsToastOpen] = useState<boolean>(false);
+
+  const searchParams = new URLSearchParams(location.search);
+  const questionId = Number(searchParams.get('question-id'));
+  const questionerName = searchParams.get('questioner-name') || '';
+
+  const handleCopyLink = async () => {
+    await copyToClipboard(window.location.href);
+
+    setIsToastOpen(true);
+  };
+
   useEffect(() => {
     if (isOpen) {
       document.body.style.overflow = 'hidden';
@@ -29,7 +47,30 @@ export const KakaoShareModal = ({ title, openGraphKey, isOpen, onClose, ...rest 
 
   return (
     <Modal isOpen={isOpen} onClose={onClose} {...rest}>
-      <KakaoShareContainer title={title} openGraphKey={openGraphKey} onClose={onClose} />
+      <section>
+        <div css={textWrapperCss}>
+          <Txt
+            color={colors.black}
+            typography="heading2"
+            fontWeight="bold"
+            css={{ width: '232px', textAlign: 'center' }}
+          >
+            {title}
+          </Txt>
+        </div>
+        <div css={buttonWrapperCss}>
+          <ShareKakaoButton openGraphKey={openGraphKey} questionId={questionId} questionerName={questionerName} />
+          <CopyLinkButton onClick={handleCopyLink} />
+        </div>
+        <Button variant="sub" onClick={onClose}>
+          닫기
+        </Button>
+        {isToastOpen && (
+          <Toast showTime={1500} onClose={() => setIsToastOpen(false)}>
+            링크를 복사했어요!
+          </Toast>
+        )}
+      </section>
     </Modal>
   );
 };
